# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Loop API",
    "version": "0.1.0",
    "description": "The Loop Autonomous Improvement Engine API. Collects signals, organizes work into issues, and dispatches AI agents with hydrated prompts.",
    "contact": {
      "name": "Loop by Dork Labs",
      "url": "https://www.looped.me"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5667",
      "description": "Local development"
    },
    {
      "url": "https://api.looped.me",
      "description": "Production"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key passed as Bearer token in the Authorization header"
      }
    },
    "schemas": {
      "Issue": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cuid2_abc123"
          },
          "number": {
            "type": "integer",
            "example": 1
          },
          "title": {
            "type": "string",
            "example": "[PostHog] error_rate: Login failures spiked"
          },
          "description": {
            "type": ["string", "null"],
            "example": "Auth failure rate exceeded 20% over the last hour."
          },
          "type": {
            "type": "string",
            "enum": ["signal", "hypothesis", "plan", "task", "monitor"],
            "example": "signal"
          },
          "status": {
            "type": "string",
            "enum": ["triage", "backlog", "todo", "in_progress", "done", "canceled"],
            "example": "triage"
          },
          "priority": {
            "type": "integer",
            "minimum": 0,
            "maximum": 4,
            "example": 2
          },
          "parentId": {
            "type": ["string", "null"],
            "example": null
          },
          "projectId": {
            "type": ["string", "null"],
            "example": null
          },
          "signalSource": {
            "type": ["string", "null"],
            "example": "posthog"
          },
          "signalPayload": {
            "type": ["object", "null"],
            "additionalProperties": {},
            "example": {
              "event": "error_rate",
              "value": 0.22
            }
          },
          "hypothesis": {
            "$ref": "#/components/schemas/HypothesisData"
          },
          "agentSessionId": {
            "type": ["string", "null"],
            "example": null
          },
          "agentSummary": {
            "type": ["string", "null"],
            "example": null
          },
          "commits": {
            "type": ["array", "null"],
            "items": {
              "$ref": "#/components/schemas/CommitRef"
            }
          },
          "pullRequests": {
            "type": ["array", "null"],
            "items": {
              "$ref": "#/components/schemas/PullRequestRef"
            }
          },
          "completedAt": {
            "type": ["string", "null"],
            "format": "date-time",
            "example": null
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-02-20T12:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-02-20T12:00:00.000Z"
          },
          "deletedAt": {
            "type": ["string", "null"],
            "format": "date-time",
            "example": null
          }
        },
        "required": [
          "id",
          "number",
          "title",
          "description",
          "type",
          "status",
          "priority",
          "parentId",
          "projectId",
          "signalSource",
          "signalPayload",
          "hypothesis",
          "agentSessionId",
          "agentSummary",
          "commits",
          "pullRequests",
          "completedAt",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "HypothesisData": {
        "type": ["object", "null"],
        "properties": {
          "statement": {
            "type": "string",
            "example": "Switching to a new auth provider will reduce login failures by 30%"
          },
          "confidence": {
            "type": "number",
            "minimum": 0,
            "maximum": 1,
            "example": 0.75
          },
          "evidence": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": ["Login failure rate up 20%", "PostHog session data shows 15% drop-off"]
          },
          "validationCriteria": {
            "type": "string",
            "example": "Login failure rate drops below 5% within 30 days"
          },
          "prediction": {
            "type": "string",
            "example": "Failure rate reduced to ~2%"
          }
        },
        "required": ["statement", "confidence", "evidence", "validationCriteria"]
      },
      "CommitRef": {
        "type": "object",
        "properties": {
          "sha": {
            "type": "string",
            "example": "abc123def456"
          },
          "message": {
            "type": "string",
            "example": "fix: correct auth token expiry"
          },
          "url": {
            "type": "string",
            "example": "https://github.com/org/repo/commit/abc123"
          },
          "author": {
            "type": "string",
            "example": "agent-claude"
          },
          "timestamp": {
            "type": "string",
            "example": "2026-02-20T10:00:00Z"
          }
        },
        "required": ["sha", "message"]
      },
      "PullRequestRef": {
        "type": "object",
        "properties": {
          "number": {
            "type": "number",
            "example": 42
          },
          "title": {
            "type": "string",
            "example": "fix: correct auth token expiry"
          },
          "url": {
            "type": "string",
            "example": "https://github.com/org/repo/pull/42"
          },
          "state": {
            "type": "string",
            "example": "merged"
          },
          "mergedAt": {
            "type": ["string", "null"],
            "example": "2026-02-20T11:00:00Z"
          }
        },
        "required": ["number", "title"]
      },
      "Label": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cuid2_abc123"
          },
          "name": {
            "type": "string",
            "example": "frontend"
          },
          "color": {
            "type": "string",
            "example": "#ff6b6b"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-02-20T12:00:00.000Z"
          },
          "deletedAt": {
            "type": ["string", "null"],
            "format": "date-time",
            "example": "2026-02-20T12:00:00.000Z"
          }
        },
        "required": ["id", "name", "color", "createdAt", "deletedAt"]
      },
      "IssueRelation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cuid2_abc123"
          },
          "type": {
            "type": "string",
            "enum": ["blocks", "blocked_by", "related", "duplicate"],
            "example": "blocks"
          },
          "issueId": {
            "type": "string",
            "example": "cuid2_abc123"
          },
          "relatedIssueId": {
            "type": "string",
            "example": "cuid2_abc123"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-02-20T12:00:00.000Z"
          }
        },
        "required": ["id", "type", "issueId", "relatedIssueId", "createdAt"]
      },
...
    },
...
  },
  "paths": {
...
    "/api/issues": {
      "get": {
        "tags": ["Issues"],
        "summary": "List issues",
        "description": "Returns a paginated list of issues, filterable by status, type, projectId, labelId, priority, and parentId.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "description": "Comma-separated list of statuses to filter by",
              "example": "todo,in_progress"
            },
            "required": false,
            "description": "Comma-separated list of statuses to filter by",
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "description": "Comma-separated list of types to filter by",
              "example": "signal,task"
            },
            "required": false,
            "description": "Comma-separated list of types to filter by",
            "name": "type",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "example": "cuid2_proj"
            },
            "required": false,
            "name": "projectId",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "example": "cuid2_label"
            },
            "required": false,
            "name": "labelId",
            "in": "query"
          },
          {
            "schema": {
              "type": ["integer", "null"],
              "minimum": 0,
              "maximum": 4,
              "example": 2
            },
            "required": false,
            "name": "priority",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "example": "cuid2_parent"
            },
            "required": false,
            "name": "parentId",
            "in": "query"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "example": 50
            },
            "required": false,
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": ["integer", "null"],
              "minimum": 0,
              "default": 0,
              "example": 0
            },
            "required": false,
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of issues",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Issue"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "example": 42
                    }
                  },
                  "required": ["data", "total"]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — invalid or missing Bearer token"
          }
        }
      },
      "post": {
        "tags": ["Issues"],
        "summary": "Create an issue",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500,
                    "example": "Investigate login failure spike"
                  },
                  "description": {
                    "type": "string",
                    "example": "Auth failure rate exceeded 20% over the last hour."
                  },
                  "type": {
                    "type": "string",
                    "enum": ["signal", "hypothesis", "plan", "task", "monitor"],
                    "example": "signal"
                  },
                  "status": {
                    "type": "string",
                    "enum": ["triage", "backlog", "todo", "in_progress", "done", "canceled"],
                    "default": "triage",
                    "example": "triage"
                  },
                  "priority": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 4,
                    "default": 0,
                    "example": 2
                  },
                  "parentId": {
                    "type": "string"
                  },
                  "projectId": {
                    "type": "string"
                  },
                  "signalSource": {
                    "type": "string",
                    "example": "posthog"
                  },
                  "signalPayload": {
                    "type": "object",
                    "additionalProperties": {}
                  },
                  "hypothesis": {
                    "$ref": "#/components/schemas/HypothesisData"
                  },
                  "labelIds": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    },
                    "example": []
                  }
                },
                "required": ["title", "type"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Issue created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Issue"
                    }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "422": {
            "description": "Validation error or parent hierarchy violation"
          }
        }
      }
    },
    "/api/issues/{id}": {
      "get": {
        "tags": ["Issues"],
        "summary": "Get issue by ID",
        "description": "Returns the issue with its parent, children, labels, and relations.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "cuid2_abc123"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Issue with related data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "allOf": [
                        {
                          "$ref": "#/components/schemas/Issue"
                        },
                        {
                          "type": "object",
                          "properties": {
                            "parent": {
                              "allOf": [
                                {
                                  "$ref": "#/components/schemas/Issue"
                                },
                                {
                                  "type": ["object", "null"]
                                }
                              ]
                            },
                            "children": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/Issue"
                              }
                            },
                            "labels": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/Label"
                              }
                            },
                            "relations": {
                              "type": "array",
                              "items": {
                                "$ref": "#/components/schemas/IssueRelation"
                              }
                            }
                          },
                          "required": ["parent", "children", "labels", "relations"]
                        }
                      ]
                    }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "404": {
            "description": "Issue not found"
          }
        }
      },
      "patch": {
        "tags": ["Issues"],
        "summary": "Update an issue",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "cuid2_abc123"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 500
                  },
                  "description": {
                    "type": "string"
                  },
                  "type": {
                    "type": "string",
                    "enum": ["signal", "hypothesis", "plan", "task", "monitor"]
                  },
                  "status": {
                    "type": "string",
                    "enum": ["triage", "backlog", "todo", "in_progress", "done", "canceled"]
                  },
                  "priority": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 4
                  },
                  "parentId": {
                    "type": ["string", "null"]
                  },
                  "projectId": {
                    "type": ["string", "null"]
                  },
                  "signalSource": {
                    "type": "string"
                  },
                  "signalPayload": {
                    "type": "object",
                    "additionalProperties": {}
                  },
                  "hypothesis": {
                    "$ref": "#/components/schemas/HypothesisData"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated issue",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/Issue"
                    }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "404": {
            "description": "Issue not found"
          },
          "422": {
            "description": "Validation error"
          }
        }
      },
      "delete": {
        "tags": ["Issues"],
        "summary": "Soft-delete an issue",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "cuid2_abc123"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "204": {
            "description": "Issue deleted"
          },
          "404": {
            "description": "Issue not found"
          }
        }
      }
    },
...
  },
...
}